+++ tests/javascript/index.php
@@ -2141,7 +2141,7 @@ function PiwikTest() {
     });
 
     test("API methods", function() {
-        expect(106);
+        expect(107);
 
         equal( typeof Piwik.addPlugin, 'function', 'addPlugin' );
         equal( typeof Piwik.addPlugin, 'function', 'addTracker' );
@@ -2231,6 +2231,7 @@ function PiwikTest() {
         equal( typeof tracker.trackPageView, 'function', 'trackPageView' );
         equal( typeof tracker.trackRequest, 'function', 'trackRequest' );
         equal( typeof tracker.queueRequest, 'function', 'queueRequest' );
+        equal( typeof tracker.disableQueueRequest, 'function', 'disableQueueRequest' );
         equal( typeof tracker.disableCookies, 'function', 'disableCookies' );
         equal( typeof tracker.deleteCookies, 'function', 'deleteCookies' );
         // content
@@ -3631,7 +3632,7 @@ if ($mysql) {
 
 
     test("tracking", function() {
-        expect(158);
+        expect(164);
 
         // Prevent Opera and HtmlUnit from performing the default action (i.e., load the href URL)
         var stopEvent = function (evt) {
@@ -3883,10 +3884,22 @@ if ($mysql) {
         ok( visitorIdStart == visitorIdEnd, "tracker.getVisitorId() same at the start and end of process");
 
         // Tracker custom request
+        var requestQueue = tracker.getRequestQueue();
+        strictEqual(true, requestQueue.enabled);
+        strictEqual(0, requestQueue.requests.length);
+
         tracker.trackRequest('myFoo=bar&baz=1');
         tracker.queueRequest('myQueue=bar&queue=1');
         tracker.queueRequest('myQueue=bar&queue=2');
         tracker.queueRequest('myQueue=bar&queue=3');
+
+        strictEqual(3, requestQueue.requests.length);
+
+        tracker.disableQueueRequest();
+        strictEqual(false, requestQueue.enabled);
+        
+        tracker.queueRequest('myQueueDisabled=bar&queue=4');
+        strictEqual(3, requestQueue.requests.length);
 
         // Custom variables
         tracker.storeCustomVariablesInCookie();
@@ -4082,7 +4095,7 @@ if ($mysql) {
             xhr.open("GET", "matomo.php?requests=" + getToken(), false);
             xhr.send(null);
             results = xhr.responseText;
-            equal( (/<span\>([0-9]+)\<\/span\>/.exec(results))[1], "40", "count tracking events" );
+            equal( (/<span\>([0-9]+)\<\/span\>/.exec(results))[1], "50", "count tracking events" );
 
             // firing callback
             ok( trackLinkCallbackFired, "trackLink() callback fired" );
@@ -4123,6 +4136,7 @@ if ($mysql) {
             ok( /myQueue=bar&queue=1/.test( results ), "queueRequest sends queued requests");
             ok( /myQueue=bar&queue=2/.test( results ), "queueRequest sends queued requests");
             ok( /myQueue=bar&queue=3/.test( results ), "queueRequest sends queued requests");
+            ok( /myQueueDisabled=bar&queue=4/.test( results ), "queueRequest sends queued requests when disabled directly");
 
             // Test Custom variables
             ok( /SaveCustomVariableCookie.*&cvar=%7B%222%22%3A%5B%22cookiename2PAGE%22%2C%22cookievalue2PAGE%22%5D%7D.*&_cvar=%7B%221%22%3A%5B%22cookiename%22%2C%22cookievalue%22%5D%2C%222%22%3A%5B%22cookiename2%22%2C%22cookievalue2%22%5D%7D/.test(results), "test custom vars are set");